requirements.txt
alt-profanity-check==0.24.0
appdirs==1.4.4
better-profanity==0.7.0
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
joblib==1.0.1
numpy==1.21.1
oauthlib==3.1.1
pyee==8.1.0
pyppeteer==0.2.5
PySocks==1.7.1
python-dotenv==0.18.0
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==0.24.0
scipy==1.7.0
six==1.16.0
threadpoolctl==2.2.0
tqdm==4.61.2
tweepy==3.10.0
urllib3==1.26.6
websockets==8.1
